requirements.txt
# ── Raspberry Pi (device side) ────────────────────────────
pyserial>=3.5
RPi.GPIO>=0.7.1       # Only works on Raspberry Pi

# ── Server side ──────────────────────────────────────────
flask>=3.0
africastalking>=1.2    # For SMS trigger sender

# ── Testing ──────────────────────────────────────────────
pytest>=7.0
